View files or clipboard without history
Designed for use with explorer send to shortcut so selected files can be sent to a viewer which doesn't add to quick access. Files can be manipulated, content cut, copied or pasted but will not hold up logon/shutdown and will never be saved.
Run with no arguments, via desktop shortcut with keyboard shortcut, to put contents of clipboard, if text, into a viewer for manipulation.